correção na tbServico

+++ laravel/database/migrations/2020_07_18_230626_create_model_services_table.php
@@ -18,7 +18,7 @@ class CreateModelServicesTable extends Migration
             $table->string('nmServico');
             $table->string('descricao')->nullable();
             $table->double('valor', 10, 2);
-            $table->integer('comissao');
+            $table->double('comissao', 6, 3);
             $table->timestamps();
         });
     }
